requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
djangorestframework==3.14.0
idna==3.4
install==1.3.5
lxml==4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.14.0
pytz==2023.3
requests==2.28.2
rich==13.3.3
sqlparse==0.4.3
urllib3==1.26.15
